requirements.txt
# Core
python-dotenv==1.0.1
supabase==2.4.0
httpx==0.27.0
requests==2.31.0

# Scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# Data
pandas==2.2.1
numpy==1.26.4

# ML
scikit-learn==1.4.1.post1
xgboost==2.0.3
tensorflow==2.16.1
keras==3.0.5

# Telegram
python-telegram-bot==21.1.1

# Kaggle
kaggle==1.6.6

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6

# Utilities
tqdm==4.66.2
tabulate==0.9.0
rich==13.7.1
